Add new code by Phuong

+++ Phone_Shop/src/main/java/com/vti/controller/AccountController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -20,6 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.vti.entity.Account;
 import com.vti.request.AccountRequest;
+import com.vti.request.AccountUpdateRequest;
 import com.vti.response.AccountResponse;
 import com.vti.service.IAccountService;
 
@@ -32,11 +34,10 @@ public class AccountController {
 	private IAccountService accountService;
 
 	/**
-	 * API getAll Account
-	 * Trả ra 1 list Account theo pagging
+	 * API getAll Account Trả ra 1 list Account theo pagging
 	 */
-	
-  @PreAuthorize("hasRole('Admin')")
+
+	@PreAuthorize("hasRole('Admin')")
 	@GetMapping
 	public ResponseEntity<?> getAllAccounts(Pageable pageable) {
 		Page<Account> entity = accountService.getAllAccounts(pageable);
@@ -46,8 +47,9 @@ public class AccountController {
 			@Override
 			public AccountResponse apply(Account account) {
 
-				AccountResponse response = new AccountResponse(account.getAccountId(), account.getUsername(), account.getFullname(), 
-						account.getEmail(), account.getGender(), account.getPhonenumber(), account.getAddress(), account.getRegisterDate());
+				AccountResponse response = new AccountResponse(account.getAccountId(), account.getUsername(),
+						account.getFullname(), account.getEmail(), account.getGender(), account.getPhonenumber(),
+						account.getAddress(), account.getRegisterDate());
 
 				return response;
 			}
@@ -59,16 +61,18 @@ public class AccountController {
 	/**
 	 * API getAccount by AccountID
 	 */
-	
-  @PreAuthorize("hasAnyRole('User','Admin')")
+
+	@PreAuthorize("hasAnyRole('User','Admin')")
 	@GetMapping(value = "/{id}")
 	public ResponseEntity<?> getAccountById(@PathVariable(name = "id") int id) {
 		Account account = accountService.getAccountById(id);
-
-		
-		AccountResponse response = new AccountResponse(account.getAccountId(), account.getUsername(), account.getFullname(), 
-				account.getEmail(), account.getGender(), account.getPhonenumber(), account.getAddress(), account.getRegisterDate());
-		return new ResponseEntity<AccountResponse>(response, HttpStatus.OK);			
+		if (account == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		AccountResponse response = new AccountResponse(account.getAccountId(), account.getUsername(),
+				account.getFullname(), account.getEmail(), account.getGender(), account.getPhonenumber(),
+				account.getAddress(), account.getRegisterDate());
+		return new ResponseEntity<AccountResponse>(response, HttpStatus.OK);
 
 	}
 
@@ -108,16 +112,30 @@ public class AccountController {
 
 		return new ResponseEntity<>("We have sent 1 email. Please check email to active account!", HttpStatus.OK);
 	}
-  
+
 	/**
 	 * API deleteAccount by AccountID
 	 */
-  @PreAuthorize("hasRole('Admin')")
+	@PreAuthorize("hasRole('Admin')")
 	@DeleteMapping(value = "/{id}")
 	public ResponseEntity<?> deleteAccount(@PathVariable(name = "id") int id) {
+		Account account = accountService.getAccountById(id);
+		if (account == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
 		accountService.deleteAccount(id);
 		return new ResponseEntity<String>("Delete successfully!", HttpStatus.OK);
 	}
-  
-  // Test merge code.
+
+	@PutMapping(value = "/{id}")
+	public ResponseEntity<?> updateAccount(@PathVariable(name = "id") int id, @RequestBody AccountUpdateRequest request){
+		Account account = accountService.getAccountById(id);
+		if (account == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		accountService.updateAccount(id, request);
+		return new ResponseEntity<String>("Update successfully!", HttpStatus.OK);
+		
+	}
 }
